requirements.txt
falcon
gunicorn
PyBambooHR
icalendar
click
redis
hiredis
pytest
mockredispy
pylint
